requirements.txt
requests
pika
hca-ingest